some updates

+++ chemvae/train_vae.py
@@ -134,26 +134,37 @@ def vectorize_data(params):
             mu.get_selfie_and_smiles_encodings_for_dataset(smiles)
         X = mu.multiple_selfies_to_hot(selfies_list, MAX_LEN,
                                        selfies_alphabet)
-        #MAX_LEN = largest_selfies_len
-        #params['MAX_LEN'] = MAX_LEN
+
         params['NCHARS'] = len(selfies_alphabet)
 
-        with open("alphabet.json", "w") as jf:
-            json.dump(selfies_alphabet, jf)
-            jf.write('\n')
+        if not os.path.isfile(params['char_file']):
+            with open(params['char_file'], "w") as jf:
+                json.dump(selfies_alphabet, jf)
+                jf.write('\n')
+        else: #union of existing alphabet with new dataset, in case of transfer learning
+            cur_chars = json.loads(open(params['char_file']).read())
+            new_alphabet = list(set(cur_chars).union(set(selfies_alphabet)))
+            with open(params['char_file'], "w") as jf:
+                json.dump(new_alphabet, jf)
+                jf.write('\n')
+
     else:
         _, _, _, smiles_list, smiles_alphabet, largest_smiles_len = \
             mu.get_selfie_and_smiles_encodings_for_dataset(smiles)
         X = mu.multiple_smile_to_hot(smiles_list, MAX_LEN,
                                      smiles_alphabet)
-        #X = mu.smiles_to_hot(smiles, MAX_LEN, params[
-        #                     'PADDING'], CHAR_INDICES, NCHARS)
-        #MAX_LEN = largest_smiles_len
-        #params['MAX_LEN'] = MAX_LEN
+
         params['NCHARS'] = len(smiles_alphabet)
-        with open("alphabet.json", "w") as jf:
-            json.dump(smiles_alphabet, jf)
-            jf.write('\n')
+        if not os.path.isfile(params['char_file']):
+            with open(params['char_file'], "w") as jf:
+                json.dump(smiles_alphabet, jf)
+                jf.write('\n')
+        else: #union of existing alphabet with new dataset, in case of transfer learning
+            cur_chars = json.loads(open(params['char_file']).read())
+            new_alphabet = list(set(cur_chars).union(set(smiles_alphabet)))
+            with open(params['char_file'], "w") as jf:
+                json.dump(new_alphabet, jf)
+                jf.write('\n')
 
     #sys.stdout.write('Total Data size ' + str(X.shape[0]))
     #sys.stdout.flush()
